Add <norecord> argument to map
<norecord> is intended to disable a mapping while recording or
executing a register.  This is especially useful for any mappings like:

    imap jk <Esc>

While recording a macro, one might be tempted to do something like

    qaij<wait a while for timeoutlen to expire>k<Esc>q

Only to find that it doesn't work when playing back the macro.

It is also useful for other plugins that turn on vim "hard-mode"
(disabling too many j/k commands in a short amount of time) or just any
plugin-mappings that depend at all on timing (accelerated-jk) or
user-interface (easymotion) semantics.