test: Prevent waiting for already current state.
Some of the UI tests screen:expect() the same screen state twice in a
row.  This ends up waiting the entire timeout period for a screen redraw
event that never happens.  So check at the start of screen:wait() that
the screen is not already in the desired end state.

This shaves a third off the total run-time of the functional tests on
FreeBSD.  Some of the tests that look for the same state twice are
mouse_spec tests, which have their timeouts raised to a quarter of a
minute for each screen:expect() in the test.