test/title: restore terminal title when quit #4063
This is a quick implementation to see the change in action.
Do note that this is a test and will see the terminal title with "Thanks
for flying Vim", after quitting vim. This is just a start and I will begin to
fix the issue. Do note with the old implementation, it seems to be
convoluted and can be simplify.
Suggestions are appreciated though.