window.c: Fix issue #7429
Link to the issue: https://github.com/neovim/neovim/issues/7429
Problem: after a help window was closed, a window was selected and its
autocommands triggered. After that, restore_snapshot was called and the focused
window changed, confusing the user.

Solution: Add a function get_snapshot_focus() that returns the window
that holds the cursor in a snapshot. This function is used in win_close
to make sure the right window is selected before any autocommand is
triggered.