[WIP] Implemented a curl wrapper. Issue: #7189
download.lua wrapps curl to be used in the spellfile.vim to surcomvent
the need of netrws download function. I'm to continue to work on the
spellfile.vim to compleate the work. I would like some input on the work
so far. Is this a good beginning to the solution of issue #7189.

Signed-off-by: mattemagikern <mansgariusson@gmail.com>