provider: run cli.js of nodejs host directly
npm shims are useless because the user cannot set node to debug mode
via --inspect-brk. This is problematic on Windows which use batchfiles and shell scripts
to compensate for not supporting shebang.

The patch uses `npm root -g` to get the absolute path of the global npm modules.
If that fails, then the user did not install neovim npm package globally.
Use that absolute path to find `neovim/bin/cli.js`, which is what the npm shim
actually runs with node. glob() is for a simple file check in case bin/ is removed
because the npm shims are ignored now.