Change g:actual_curbuf variable to use set_internal_number_var
Used when evaluating %{} blocks in the statusline, the variable g:actual_curbuf
is set so the expressions can see both the statusline-local current buffer, as
well as the global, actual current buffer. However, because it used
set_internal_string_var, actually using this variable in vimscript required
coercing it to a number variable before it could be used in any API calls, to
avoid an "Expected Buffer" error. Now the g:actual_curbuf variable can be used
directly.