Add statusline window variables v:stl_winid and v:actual_curwinid
`v:stl_winid` is set during evaluation of `%!` statuslines, and holds the
ID of the window the statusline is currently being drawn for.

`v:actual_curwinid` is set during evaluation of `%{}` statusline expressions,
and holds the ID of the actual current window (the one the user is editing in),
rather than the window the statusline is currently being drawn for.

This allows `%!` and `%{}` in statuslines to have rough parity in their
capabilities and makes certain statusline highlighting situations much easier
to configure.