unittests: Attempt to make out of bound pointer crash
The idea is to allocate a whole memory page (or a number of whole memory pages)
and place value just at the end of the allocated region, so any out of bounds
access (or, actually, any access past the end of the region, not before) will
inevitably occur in the region allocator is not going to claim and cause SEGV.

Strangely, attempt to do that only works only if I allocate too much at once,
and supposedly only because I allocated too much memory and not because there is
any out-of-bounds access involved.

[ci skip] *Tests are known to crash at this commit.* This commit is supposed to
be a fix for test which is supposed to crash, but does not, *not* for a crashy
application. Crash was not yet investigated fully, but is already known to
really be for an unrelated reason: test known to be fine turned into
pagealloc-using test is crashing as well. Also 48 GiB is above the amount of
memory my system has while 32 GiB is not (if swap is considered, but not the
fact that much of the non-swap memory is used and not only by FS caches). It
could be fine with overcommit: it is not like I actually accessed much of the
memory in the allocated address space, but apparently overcommitting that much
does not sound fine for the kernel, tcc running a simple C app can neither
allocate 48 GiB.