tui: focus reporting: use xterm escape sequence for urxvt 9.23+
urxvt supports the xterm sequence for focus reporting in master / CVS
since 2016 [1].

This has not been released yet, but is expected to be in 9.23.

In 3d0ee17 a special case for rxvt was added, which requires a custom
urxvt extension, see #7578.

Since urxvt 9.23 is not released still, and extensions for handling this
manually appear to be in use, this sends both escape sequences for rxvt.

Fixes https://github.com/neovim/neovim/issues/8695.

1: https://github.com/exg/rxvt-unicode/commit/75264fa544529a21ff79d3da26eb3130c43444a5#diff-2c8f5590ce4fa7495edcf7572c89c77b