eval: handle profiling and verbose in system()
The code to handle verbose and profiling seems to have been lost on a
refactoring of `evalbuf.c` (in vim codebase) into `eval.c` (neovim).

Vim's code calls `call_shell` directly from `get_system_output_as_rettv`()
whereas in neovim this function has been rewritten to not call `call_shell` but to call
`os_system` via `do_os_system`, loosing the support for profiling and verbose.

Changing the code to call `call_shell` from `get_system_output_as_rettv`
seems to be too complicated to be worth it on the current version of the
code. It seems to be simpler to just re-add the code to handling the
missing bits on its current version.