eval: remove extra call to smsg
I still do not know the code base enough to know why the way it was
written before didn't work.

In this commit I changed it to mirror the
way other functions in the same file print verbose information (calling
`verbose_enter_scroll` / `verbose_leave_scroll`) and it worked as expected.

I added a msg_puts to separate the error message from the the
command output. I added a double '\n' just to make it match the current
output for the '!' counterpart (i.e. test `:4verb !echo hi`).

This commit also changes some white spaces in the functional test to
match the actual output of the call.