Add "Transitioning from Vim" section to README
After installing neovim, I tried following the instructions of several
outdated blog posts to transition from vim to nvim.

This commit adds a prominent link to the README to the docs on this
subject, right after installation instructions. This makes sense to me
because most new neovim users will come from vim (at least for now).

Also, places the "features" section higher in the README, as it pertains
to more people than the project layout section, which is for those who
wish to contribute.