RPC: log invalid message
Usually we dump the message contents only for DEBUG-level logging. But
INVALID message is unusual, so we should always dump it (helps with bug
reports, for example).

Caveat: the message contents of an invalid message aren't actually that
useful because msgpack-c probably failed to parse it.  So in practice,
this change only gives us the message id and type.