Moved to a much clearner impl of cbs
Wow, can't believe that this actually ran after I refactored so much of
how callbacks are handled...

but I'm glad I refactored it. Should be alot easier to add more
callbacks in the future, as well as change the behavior for
configuration and options.

Also moved stuff to quickfix/loclist and made the first "configuration"
options for those. I'll expose those later.